[PATCH] Wasted memory use storing path copies in node tree When creating the node tree, every node stores a copy of its full path table. e.g. for node("admin.network.wireless"), node.path = { "admin", "network", "wireless" }
This value is not used anywhere, and likely may be from before the addition of the treecache lookup table? In any instance, I've searched high and low and see nothing ever referencing any node's path via the path member. It 
eats a good chunk of memory though and as such I believe it should be removed.

I've tested every page in the admin-full module after removing it and all seem to function properly.


git-svn-id: http://svn.luci.subsignal.org/luci/trunk@7356 ab181a69-ba2e-0410-a84d-ff88ab4c47bc